requirements.txt
pandas==1.1.2
numpy==1.19.2
pycaret==2.2.2
streamlit==0.71.0
Pillow==8.0.1
